Use TimeLib.h instead of Time.h
The new version of avr-libc included with Arduino IDE 1.6.10/Arduino AVR
Boards 1.6.12 has a file named time.h. This causes that file to be
included instead of the intended file on case insensitive operating
systems, resulting in compile failure. This is also an issue with
non-AVR boards. Recent versions of the Time library have a file named
TimeLib.h to solve this issue.